+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Web Puzzlers Triggers Changelog
 
+## 1.0.3 - 2022-05-12
+
+### Changed
+- Changed EntryDraft, EntryRevision and EntrySlug handles
+
+### Added
+- Commerce triggers
+
 ## 1.0.2 - 2022-05-10
 
 ### Fixed